More developer-centric superbuild
I have been meaning to get to this for a while, and having a few new
machines to get set up was a good motivator. This currently builds
ParaView the way Tomviz needs it, and then builds Tomviz pointing it at
the ParaView build tree (Tomviz cannot use an install tree at this
time). The TBB backend is disabled as it was causing a CMake error for
me locally, it also isn't necessarily needed for a development build.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>